Fix item-from-map-consumption:
Makes the quantity parameter of the map/item functions a reference, so the number of consumed items/charges can be removed from it and the caller known about the remaining ones.

Make the quantity parameter of the use_amount functions a long, for consistency and simpler usage (calculated from the same source).

Some functions need to use a dummy variable to provide a reference when calling the changed functions.
Some functions need a temporary variable as they originally use an int, but must provide a reference to long when calling further functions. They should get the long directly as parameter, but that would change the function signature and can wait for later.

+++ src/item.cpp
@@ -4140,7 +4140,7 @@ int item::amount_of(const itype_id &it, bool used_as_tool) const
     return count;
 }
 
-bool item::use_amount(const itype_id &it, int &quantity, bool use_container, std::list<item> &used)
+bool item::use_amount(const itype_id &it, long &quantity, bool use_container, std::list<item> &used)
 {
     // First, check contents
     bool used_item_contents = false;
